requirements.txt
speedtest-cli
rich
wheel
